requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
pandas>=2.0.0
matplotlib>=3.5.0
numpy>=1.20.0
python-dotenv>=1.0.0
scipy>=1.7.0
statsmodels>=0.13.0
